Explicitly set inventory item value to uint16_t

+++ src/inventory.hpp
@@ -24,9 +24,9 @@
 
 namespace inventory
 {
-    void player_getitem(std::string const& p_name, unsigned int val);
-    void player_useitem(std::string const& p_name, unsigned int val);
-    unsigned int return_item_n(std::string const& p_name);
+    void player_getitem(std::string const& p_name, uint16_t val);
+    void player_useitem(std::string const& p_name, uint16_t val);
+    uint16_t return_item_n(std::string const& p_name);
     void display_screen();
 }
